refactor classes to javax space
